+++ common/src/main/java/com/macuguita/woodworks/block/ResizableBeamBlock.java
@@ -95,7 +95,13 @@ public class ResizableBeamBlock extends Block implements Waterloggable {
 
 	private final VoxelShape[][] radiusToFacingsShape;
 
+	private final boolean strippable;
+
 	public ResizableBeamBlock(Settings settings) {
+		this(settings, true);
+	}
+
+	public ResizableBeamBlock(Settings settings, boolean strippable) {
 		super(settings);
 		this.radiusToFacingsShape = this.generateRadiusToFacingsShapeMap();
 		this.setDefaultState(this.getStateManager().getDefaultState()
@@ -107,6 +113,7 @@ public class ResizableBeamBlock extends Block implements Waterloggable {
 				.with(UP, false)
 				.with(DOWN, false)
 				.with(WATERLOGGED, false));
+		this.strippable = strippable;
 	}
 
 	@Override
@@ -195,7 +202,7 @@ public class ResizableBeamBlock extends Block implements Waterloggable {
 	protected ActionResult onUse(BlockState state, World world, BlockPos pos, PlayerEntity player, BlockHitResult hit) {
 		Hand hand = player.getActiveHand();
 		ItemStack stack = player.getStackInHand(hand);
-		if (stack.getItem() instanceof AxeItem) {
+		if (stack.getItem() instanceof AxeItem && strippable) {
 			Block strippedBlock = STRIPPED_BEAM_BLOCKS.get(this);
 			if (strippedBlock != null) {
 				if (!player.getAbilities().creativeMode) stack.damage(1, player, LivingEntity.getSlotForHand(hand));